// frontend/src/backend.d.ts
export interface BannerMessage {
    order: bigint;
    enabled: boolean;
    message: string;
}